Allow treating callable-typed parameters as being called at call sites
Summary:
# Problem
Sometimes at a call site, we wish to treat the callable-typed parameters as
being called in the taint analysis. However this is incompatible with the
higher order call graph building, which would always remove such parameters if
it manages to create parameterized callables.

# Solution
This diff allows the higher order call graph building to still keep some
callable-typed parameters that are hardcoded. In the follow-ups we will allow
users to specify them, instead of hardcoding.

Reviewed By: arthaud

Differential Revision: D74196942

fbshipit-source-id: 0f4e9855ed385a5bf770cbb53680ff6ca5e0bd22