Preserve the location of AST nodes created from string annotation expansion
Summary:
In some cases, we expand a string literal into a parsed AST.
For instance, in `x = typing.cast("typing.Union[int, str]", y)`, the "typing.Union[int, str]" is parsed into a type annotation.

In order to fix a bug with go-to-definition (see D38788553), we discard the location of the parsed AST and use the location of the string for any subexpression.
This is causing issues for the uniqueness of expressions (see context in D74729695) for Pysa.

This diff proposes to add an optional to either preserve the original location of the string (current behavior) or use the location from the parsed AST.
We can then turn this option on for Pysa only.

Reviewed By: tianhan0

Differential Revision: D74735202

fbshipit-source-id: a06f98ae32af6b9577cc0cc078ba2a490c43ff13