Fix uniqueness of self implicit type var origins
Summary:
# Context
See original context in D73845705

# Short term Objective
We are trying to build a type that would uniquely identify a given expression.
The idea would be to use the `Origin.t` of the expression when present, or the `Location.t` otherwise (which would mean the expression is original, from the source code)

# Problem
There are still many cases where we create multiple expressions with the same origin+location.
One of these case is due to self implicit type variables, which we create for each class in a module.
Those are currently all assigned the invalid location (-1) hence if a module has 2 classes, it will get two self type variables at the same location (-1) with the same origin.

To fix the problem, let's add the self variable name in the origin.

Reviewed By: tianhan0

Differential Revision: D74729695

fbshipit-source-id: b541f45d3d9ac0c05c393c8473fb2d365873d189