Use different origins for implicit __str__ and __repr__ call
Summary:
There are some edge cases where we create both `x.__str__()` and `x.__repr__()` for the same `x` expression.

This happens for instance in `if isinstance(x, str) and str(x) == "foo"`.
This is turned into `assert(isinstance(x, str) and str(x) == "foo")` and `assert(not(isinstance(x, str) and str(x) == "foo"))`. In one case, we are able to refine that `x` is `str` and therefore `str(x)` calls `x.__str__()`, and in the other branch, it calls `x.__repr__()`.

Since we want origins to be unique (see D74729695), let's use different origins for `__str__` and `__repr__` calls.

Reviewed By: tianhan0

Differential Revision: D74803644

fbshipit-source-id: dbb11f7cb4403eb76471c43c6e5875ee49929cbb