Store type preconditions for statements that lead to the unreachable state
Summary:
We currently have a bug in Pysa where we are missing type information in some statements.
It looks like local type annotations are not saved when the statement leads to the unreachable state.
For instance, this can happen in `if isinstance(x, int):` when Pyre is smart enough to know that `x` will always be an int.

Reviewed By: tianhan0

Differential Revision: D75079648

fbshipit-source-id: 1a78d426326b7d438e26d82cf4957537ac27a348