Fix non-unique expression identifier for augmented assigns with properties
Summary:
# Problem
We translate statements such as `self.foo &= 0` into `self.foo = self.foo & 0`.
If `foo` is a property, we will actually have different callees for the `self.foo` on the left hand side (calling the setter) and the right hand side (calling the getter).
However, if we use the expression identifier to index the call graph (which we will do in the next diff), then we would end up with the callees being merged since there is a single `self.foo` expression.

# Solution
Create two different origins for the left hand side and right hand side of augmented assignments.

Reviewed By: tianhan0

Differential Revision: D75136436

fbshipit-source-id: 2650a37263c81644e4abdcd375f31a55c6dac273