Use the expression identifier to index the call graph
Summary:
In D74727882, we introduce a type that can uniquely represent an expression.

Right now, our call graph uses a mix of the location of the expression, and a "expression identifier" pretty loosely defined. It is sometimes the whole expression, or sometimes just the accessed attribute name.
Let's replace those heuristic by the expression identifier introduced in the previous diff.

Reviewed By: tianhan0

Differential Revision: D75061940

fbshipit-source-id: f171b348ff4aa8c6e18aa1727cb03320a6376a50