add "enable_strict_any_check" flag
Summary:
Pyrefly is going to typecheck all function bodies, but it won't by default complain about "Any".

This diff brings pyre closer to the pyrefly behavior. The plan is:
1. Right now we'll switch *every single file* over to `# pyre-strict`. This will cause it to typecheck all function bodies. (Previously, about half of fbcode functions didn't have any type annotations and weren't in strict files, hence weren't being typechecked).
2. But first we'll change the default behavior of `# pyre-strict` so it no longer complains about Any. That was too much.
3. There are some folks in IG who continue to care about Any. For them, we'll add a flag in instagram_server/distillery/.pyre_configuration_local `"enable_strict_any_check"` so that, for them, strict mode will continue to complain about Any.

Why are we doing it now rather than waiting for pyrefly? (1) by doing it now, the delta we see when switching to pyrefly will be much smaller; we're front-loading a lot of the changes to fbcode. (2) I'm impatient to use AI to add missing python annotations, but I want to keep "light up typechecking in function bodies for the first time" completely separate from "add python type annotations".

This diff merely adds the `enable_strict_any_check` flag to pyre (frontend and backend) and plumbs it all the way through. It doesn't yet change any behavior; that will come in the next diff.

This diff is based on the precedent of D60078685, which added "enable_strict_override_check": that flag caused strict mode to check overrides, which strict mode doesn't normally do.

Reviewed By: stroxler

Differential Revision: D75254948

fbshipit-source-id: 067ab093498faef663b9a6f95ae9523680f0c025