change behavior on "enable_strict_any_check" flag
Summary:
The previous diff summary explains the goal, rationale, and workflow. The essence is we're adding a new flag `"enable_strict_any_check"`: (1) by default it's false, and strict mode doesn't complain about Any; (2) but if .pyre_configuration.local sets it to true, then strict mode will complain about Any.

This diff changes the behavior of the typechecker to respect that flag. The change in behavior is solely in this function:
 {F1978536928}

(This diff follows the precedent of D60085379, which changed the behavior of the typechecker to respect the flag `"enable_strict_override_check"`)

Reviewed By: stroxler

Differential Revision: D75255383

fbshipit-source-id: f4e2718957255c2a83c3d8151fe9725d26b379b5