Fix inconsistencies with redirect expressions and call graph building
Summary:
# Context
As described in D73845705, we perform a large set of AST transformations, to imitate the python runtime. For instance, we translate `a + b` into `a.__add__(b)`.
Pysa performs most of these transformations in the function `redirect_expression` of the call graph module, which is called before analazing any expression.

# Problem
`redirect_expression` will produce different ASTs depending on the current typing context. For instance, `str(x)` might resolve into `x.__str__()` if that function exists, otherwise it will resolve into `x.__repr__()` - which always exists since `object` defines `__repr__`.
The problem is that the call graph building, forward and backward analysis have slighly different logic to propagate the pyre context when visiting generators.
A generator is an expression such as `[f(x) for x in l if cond(x)]`. To get the correct type information, we need to tell pyre to bind `x` to `l.__iter__().__next__()` in the type context.
The call graph building does not currently do that correctly, because it uses the legacy AST visitor. It updates the `pyre_context` in the state, but doesn't unset it, so having nested generators lead to incorrect results.

# Solution
To solve this problem, this diff makes the following changes:

* `redirect_expression` is rewritten to be recursive, so it only needs to be called once in `forward_statement` of call graph building, forward and backward analysis. It uses `Expresison.Mapper` under the hood, which gives us more control. More specifically, it allows us to properly update the pyre context when visiting generators.
* `redirect_expression` is renamed `preprocess_expression` since it reflects better what the function does.
* Call graph building is rewritten to use `Expression.Folder` instead of the legacy visitor. This also fixes the problem with the inconsistent pyre context in generators.

Reviewed By: tianhan0

Differential Revision: D75452799

fbshipit-source-id: 7e30e674a48d17c499334b421f5f39ef20c42342