Update pyre client command line options for higher order functions
Summary:
This diff:
1. removes the command line option for inlining decorators, since the backend
no longer supports this.
2. adds the command line options for configuring the building of higher order
call graphs.

This diff should be ok to land as is, because the changes in the CI testing
scripts are to remove specifying the inlining of decorators, which is not
supported at this moment. Hence removing specifying `--inline-decorators`
should not change the expected results of CI tests.

Reviewed By: arthaud

Differential Revision: D75258802

fbshipit-source-id: cc0d9b5a56f6c749b6c79628f59e1757ae3cb1f8