Cache type of expressions required for preprocessing
Summary:
We preprocess statements and expressions to properly handle dunder methods. For instance, `a + b` is turned into `a.__add__(b)`.
During this preprocessing, we sometimes need the type of an expression. Let's use the `TypeOfExpressionSharedMemory` so types of expression are only requested once and stored in the cache.
This should improve performance.

Reviewed By: tianhan0

Differential Revision: D76426943

fbshipit-source-id: 61bd2f489a05314fdcac7e6f14a31d2e3ab10123