Fix a crash due to visiting the right hand side of chained assignments multiple times
Summary:
We treat `e1 = e2 = e3` as `e1 = e3; e2 = e3`. This can lead to inconsistencies when building the call graph, in cases `e3` resolves to different callables in the first statement and the second statement.
Most notably, this happens when statements have some implicit dependencies, such as `x = y = x.method()`.

To prevent this issue, we skip revisiting `e3` when the assignment is marked with the `ChainedAssign` origin.
However, in some cases, the assignment is turned into a call, for instance when the left hand side is a subscript: `d[a] = b` is turned into `d.__setitem__(a, b)`. We need to skip visiting `b` if this is part of a chained assignment.

To make this change, we need to use `Expression.Origin.t` as the origin of assignments, instead of using its own type `Assign.Origin.t`.

Reviewed By: tianhan0

Differential Revision: D76597301

fbshipit-source-id: b9ded8ba2352e25be5bab2b53eb580fb07ca2b53