Do not delocalize callables names when exporting them
Summary:
See D76745214 for more context.

We are delocalizing callable names when we export them, turning `$local_foo.bar` into `foo.bar`. While we shouldn't create localized callable names in the first place, this is making it harder to even detect the problem, since the taint output will have the correct callable names.

Let's remove this, and export the callable name as is it in memory.

Reviewed By: tianhan0

Differential Revision: D76745822

fbshipit-source-id: 2f5c69f0a8b87121c9ae41a6471fd70ca53d99ea