Wrap the callable function and method type in a module
Summary:
We recently found a bug where we create invalid callable names, using localized names (e.g, `$local_`). This is leading to duplicate callable names, and false negatives.
Right now, the creation of callable is spread all over the codebase and has no checks, so it makes it hard to find the root cause.
This diff proposes to force the creation of callables through one of the `Target.create_xxx` functions.
In future diffs, we will have some sanity checks to prevent the creation of invalid callables.

Reviewed By: tianhan0

Differential Revision: D76745214

fbshipit-source-id: 41fd8fe021980aa7dd85b87d44162ea598aa8f03