Add a test demonstrating a bug in the call graph building
Summary:
We sometimes get confused whether a callable is a function or method. The way it is decided is very error-prone, and relies on type information.

This diff adds a test demonstrating a bug due to this confusion.

This can lead to false negatives since we will have different versions of callables, one with a proper model, and one without model.

Reviewed By: tianhan0

Differential Revision: D76811060

fbshipit-source-id: e2de4efb3d4d646020a552f7a32aa2f8558f913a