Implement a way to attach a breadcrumb to all variables in the scope
Summary:
This diff implements a new kind of model called `AddFeatureToState`. The syntax is as follows:

```
AddFeatureToState(Via[foo])
def my_module.bar(): ...
```

This will add the feature `foo` to all local variables in the state when the function `bar` is called.
This is useful to track calls that happen outside the data flow, for instance:

```
x = source()
if bar():
  sink(x)
```

Technically, `bar` is not involved in the data flow. Sanitizing `bar` would have no effect.
As a work-around, using `AddFeatureToState` allows to quickly find issues where that call happens (or does NOT happen).

Keep in mind that this has many downsides, as demonstrated in the integration test.

Reviewed By: tianhan0

Differential Revision: D77313163

fbshipit-source-id: 5ffda2e8fa0206805d59f28817e819fef5ac3823