Properly handle calls to a mix of functions and bounded methods
Summary:
Right now, we don't properly handle calls that can dispatch to a function and to a bound method (with `implicit_receiver=true`). This is because the taint representation for the implicit argument is an enum.
We can turn it into a record instead to handle a mix of functions and bounded methods.

Reviewed By: tianhan0

Differential Revision: D77666758

fbshipit-source-id: 7b3593dee4bbf3d3ba24d2195afa432b7b0f5d8d