Store shim targets in the call graph and handle them in the analysis
Summary:
# Context
See context in D77446711

# This diff
This diff stores the shims in the call graph (the set of callees and the argument mapping), so it can be used during the taint analysis.
It also updates the taint analysis (forward and backward) to analyze those shims properly.

Unfortunately, this introduces a few false negatives and false positives, because the analysis of shims happens after we already analyzed the arguments of the original call and lost some of the information. If this is important, we can revisit this.

Reviewed By: tianhan0

Differential Revision: D77597038

fbshipit-source-id: 1dce138a729dd901b2f073e5b814daa7f612dc12