Add a representation for shims at a call site
Summary:
# Context
"shimming" is a way to treat a function call as another function call, potentially remapping parameters.
For instance, we shim `functools.partial(f, a)` into `f(a)` in order to catch flows to `f`, even if `functools.partials` does not actually call `f` right away.

Shimming in Pysa is currently implemented as AST transformations, i.e we literally transform the code before the analysis.
Wile this worked until now, this probably won't work with Pyrefly, since we won't be able to type check arbitrary AST expressions (most likely).

# Solution
To fix this problem, we want to re-implement shims without AST transformations.
For that, we plan to store shim targets in the call graph, and analyze those during the forward and backward analysis.

# This diff
This diff is the first of the stack. It implements a concept of argument (and callee) mapping. This describes how the callee and the arguments of the artificial/shimmed call (e.g, `f(a)` for the example above) is constructed from the original expression. The uses a type `Target.t` which allows to grab a specific argument or the callee of the original call and apply some light transformation (get the base attribute, get a list element, etc).

We then change `specialCallResolution.ml` to emit this argument mapping. For now, we introduce a function that performs the AST transformation, so this diff should be a no-op.

Note that this representation is not meant to be specified by users directly. I believe we will need another abstraction layer (DSL) for the user. The main reason is that the current representation requires to describe the origin of the artificial arguments, but the number of arguments might not be known statically. For instance, `multiprocessing.Process(target, args=[a, b, c])` is shimmed and expands into a dynamic number of parameters, i.e `target(a, b, c)`.

Reviewed By: tianhan0

Differential Revision: D77446711

fbshipit-source-id: 9b89e090b5b33c845b2b07e81b42f23966b2083a