Preserve keys on dict.update with dict argument
Summary:
We can be more precise when analyzing `dict.update` calls, as shown in the previous diff D78334609.
The main problem with `dict.update` is that it also accepts lists of `(key, value)` pairs. In that case, we would need to transform the list into a dict during the analysis, which we don't currently do.
For now, we can hardcode the most common case, when `dict.update` is called with a `dict` argument, and perform the more precise analysis.
The downside is that we now need more type information during the taint analysis.

Reviewed By: tianhan0

Differential Revision: D78336053

fbshipit-source-id: 636ef0d45e31b02e302e3c8634f23bf9658fb14a