Add scaffolding to support pyrefly as input
Summary:
See context in D78022303.

Now that we added an option to provide pyrefly results (D78406202), this diff adds scaffolding to allow us to use the pyrefly results during the taint analysis.
This is done by implementing a `Interprocedural.PyrePysaApi` module just wraps the old `Analysis.PyrePysaEnvironment` module, which provides information from Pyre 1 (OCaml).
In future diffs, we will implement an API for the Pyrefly results.

Reviewed By: tianhan0

Differential Revision: D78415750

fbshipit-source-id: 7d764b4b96da869a32550b24a51f9abae2016a85