Parse the file format written by pyrefly
Summary:
See context in D78022303.

This diff implements the parsing of the files exported by pyrefly for pysa.

The "project" file, i.e pyrefly.pysa.json is basically a map from module id to some module information (id, name, path, info path).
Each "module info" file represents information from pyrefly about a module. It contains the module id, name, path, as well as a map from location to type, and location to go to definition.

Note that we don't implement the parsing of go to definitions yet, because we don't need it yet to go further.

Reviewed By: stroxler

Differential Revision: D78488183

fbshipit-source-id: 8162dd81d30690125572d21b93645049e6de21c1