Refactoring to handle source paths with the same module name
Summary:
# Context
Pyre made the early assumption that a given module name could only map to one file. For instance, `my_module.main` could map to `~/project/my_module/main.py`.
However, this is generally not true when analyzing a project with multiple root directories.
Pyrefly removed that assumption, which allows it to analyze source files that would map to the same module name. This is useful when type checking a source file that also has a stub, for instance `a.py` and `a.pyi`.

# Problem
Since Pysa is based on Pyre, it suffers from the same (wrong) assumption that one module name can only map to one source file.
Since we are working on using Pyrefly's result to power Pysa's analysis, we want to handle that problem.

To solve the problem, we want to introduce a unique identifier for a given module name + path.
However, Pysa heavily relies on module names to build fully qualifier names, which are user facing in our results. This means we can't use integer ids as unique identifiers.

This diff introduces a new scheme for module names used when importing Pyrefly results.
When a module name maps to a single source file, we will simply use the module name as the unique module identifier.
When a module name maps to multiple source files, we forge a module identifier using the last elements of the path and the module name. For instance, this could be `a.pyi#b.a` and `a.py#b.a` if we have two files `b/a.pyi` and `b/a.py`.

Reviewed By: tianhan0

Differential Revision: D78754512

fbshipit-source-id: 14ac77e20f0783834b98aec9307abaa729868c82