Rename Qualifier into ModuleQualifier and change storage
Summary:
In a future diff, we will introduce another type to represent a fully qualified name.
To avoid name conflict, let's rename Qualifier into ModuleQualifier.

This diff also changes the implementation to store a Reference.t, since this is
what will be used by Pysa down the line. We can change this in the future.

Reviewed By: tianhan0

Differential Revision: D78830897

fbshipit-source-id: ecc1212781b1f49343add122ffcf33717b26a622