Collect definitions and classes with fully qualified names
Summary:
TL, DR: This diff builds on top of D78754512 to build fully qualified names for each class and function/method, that are unique across the analyzed code.

# Context
Pyre 's design uses a qualification step that gives "fully qualified names" to all classes, function and methods. This is usually a dotted path containing the module name and a "symbolic" path to the definition in the callable. For instance, `my.module.MyClass.foo`.
The fully qualified name is then often used as a key to information (for instance, mapping to a function type signature).

However, the names that are built are not actually unique. For instance, two consecutive definitions with the same name would have the same fully qualified name. Pyre somehow handle this (at least in some part of the codebase) by storing a list of definition for a given fully qualified name.

Pysa, on the other end, needs each fully qualified name to represent a unique definition or class. When there are multiple definitions with the same name, we simply pick the last one, and ignore the other ones. This can lead to false negatives.

We are currently working on using Pyrefly's result to power Pysa's analysis.
Pyrefly does not rely on fully qualified names in its design, so it doesn't suffer from the same problem.

However, Pysa still needs fully qualified name, because our names are user facing. They are considered part of the identity of an issue. Changing the naming scheme of functions would make us lose a lot of triaging information.

# Solution
This diff builds a fully qualified name that is unique for each class, function and method.
To do that, it uses the module qualifier built in D78754512, which uniquely represents a module.
Then, for each AST, we traverse all definitions, preserving a scope with naming information.
To handle conlicts within a given scope, we add a suffix to the name, such as `$1`, `$2`, etc.
To handle conflicts across modules, we sometimes need to add a specific separator between the module qualifier and the local qualified name.
In most cases, the fully qualified names will be the same as before, i.e `my.module.MyClass.foo`.
In case of conflicts, it can turn into something like `some/path.py:my.module#MyClass.foo$1`.

Reviewed By: tianhan0

Differential Revision: D78832811

fbshipit-source-id: 09e34ab5390c4c9372a1a846be96619c3e3b7476