Allow `ReturnShimCallees` to co-exist with other `ExpressionCallees`
Summary:
# Problem
See the newly added unit test. Since `ReturnShimCallees` may be attached to
*any* type of expression, it is possible for an expression to be attached with
both `ReturnShimCallees` and `ExpressionCallees`. This is currently not allowed.

# Solution 1 (V1)
We can allow an expression to have both `ReturnShimCallees` and
`ExpressionCallees`. Another reason is that, `ReturnShimCallees` is attached to
return statements (as opposed to *expressions*). Since the call graph is a
function from expression types and locations to callees, we do need a way to
allow attaching callees to statements.

Technically, `DefineCallees` is also attached to define statements, instead of
expressions.

# Solution 2 (Latest version)
We can use the location of the return statement (instead of the return
expression). There should be no conflict because there is unique callees on the
return statement location.

Reviewed By: arthaud

Differential Revision: D78828106

fbshipit-source-id: 1c1ea3fcb367c512f607dbbf149b9043ae31ecc3