Support specifying partial flows in taint configuration files
Summary:
For context, we'd like Pysa to support partial flows, which is implemented in
SAPP post-processing.

To make this work, we only need the `taint-metadata.json` file (which is in the
Pysa outputs) to contain partial flow configurations, which however needs to
come from users. This diff allows users to specify the partial flow
configurations in taint configuration files, which can be
1. parsed (this diff) and
2. written into `taint-metadata.json` files (in the follow-up diffs).

Reviewed By: arthaud

Differential Revision: D79380674

fbshipit-source-id: c4034261085673423956177d8919a038742f9e38