Fix missing log lines
Summary:
When running the analysis on small codes, we sometimes miss the last few lines of log.

This is a design problem: the `pyre` command is implemented in python, which calls the ocaml `pyre.bin` program, and stores stderr in a file. We then have a manual implementation of `tail` that looks for new lines and prints those. In some cases, the python process exits before we can see the last few lines in the stderr file.

As a workaround, let's just wait 0.01 second in the python process, for our `tail` implementation to see the last log lines. It seems to work consistenctly:

Reviewed By: tianhan0

Differential Revision: D80159207

fbshipit-source-id: b6b2839f6e74d0b660df0d2c1f8b18b0f5043054