Avoid transforming `class` models during parsing
Summary:
We currently support modeling classes, for instance:
```
SkipOverrides
class Foo: ...
```

The way it is implemented is that the statement is transformed into:
```
SkipOverrides
def Foo.method1(): ...
SkipOverrides
def Foo.method2(): ...
SkipOverrides
def Foo.method3(): ...
```

This is done during parsing, which is less than ideal, because it means we need class information (all methods) while parsing models. This should be done in a separate step.

This diff moves that logic at the model instantiation step.
This diff also adds an error when the class doesn't exist, instead of silently ignoring the model.

Reviewed By: tianhan0

Differential Revision: D80184306

fbshipit-source-id: 6ae0425c5120f418014aacf0658aef428d6a9a90