Remove the type in the return value of resolve_global
Summary:
During model parsing, we call pyre to check if a reference (e.g, `foo.bar`) points to a valid module/class/callable/attribute. That logic is implemented in `PyrePysaEnvironment.ReadOnly.resolve_qualified_name_to_global`.

To migrate to pyrefly, we will need to provide a similar API. However, the current API returns a lot of information that is not needed, and would be more expensive to get for Pyrefly. For instance, when querying for an attribute, it returns its type, but never uses it.

Let's remove `Type.t` from `Global.t` when it's not needed.
We also change `Global.t` to explicitly distinguish between function/methods vs regular attributes.

Reviewed By: tianhan0

Differential Revision: D80260665

fbshipit-source-id: d4300715d83401b3904c221ac1e31971d59ee7da