Add implicit top level modules
Summary:
# Context
For model parsing, we need a way to perform lookups of fully qualified names, such as `django.http.HttpRequest.POST`. For this, we will probably need to lookup the module first, using each possible prefixes of the symbol (`django.http.HttpRequest`, django.http`, `django`).

When none of the prefixes are valid module names, then we can give an error to the user that `django` is not in the current environment, and probably needs to be installed, or that something is wrong in the search path.

# Problem
In some cases (although rare), Pyrefly won't create intermediate modules (e.g, `a` in `a.b.c`) unless there is a `a.py` or `a/__init__.py` source file. This could lead us to wrongly reporting modules as missing in the environment when parsing models.

This diff creates implicit top level modules for all intermediate modules, so it can be found during model parsing.

Reviewed By: tianhan0

Differential Revision: D80266643

fbshipit-source-id: 896539b8da88eebcc6d57dd5c629aae9f0415b0a