Move the special logic for properties in resolve_qualified_name_to_global
Summary:
We currently have special logic in modelParser.ml to find property getters and property setters. This logic is specific to Pyre1, therefore let's move it in `resolve_qualified_name_to_global` in pyrePysaEnvironment.ml

We also have to change the function return type with more information so we can reconstruct the callable name.

Reviewed By: tianhan0

Differential Revision: D80550247

fbshipit-source-id: 4d40b2f43e1b34c225e20ef6de9963595c9b26ae