Prototype of `resolve_qualified_name_to_global` for pyrefly
Summary:
This diff implements a prototype of the `resolve_qualified_name_to_global` API when using Pyrefly.

For now, we just perform lookups for a module, callable or class with the exact name. In the future, we might need to perform some fuzzy matching logic when dealing with non-unique module names.

Reviewed By: tianhan0

Differential Revision: D80639323

fbshipit-source-id: 42b1dbf4aae3067247c477f9ad64b45c14f06fc1