Replace the `ReadOnly.class_summaries` API by `ReadOnly.class_method_signatures`
Summary:
# Context
During model parsing, we currently use the `PyrePysaApi.ReadOnly.class_summaries` API which returns the whole AST definition of a class.
In practice, we never need the whole AST, and only want to get the list of method names with their signatures (parameters + decorators).

# Problem
When using Pyrefly, we don't currently store the whole AST node of the `class X: ...` statement, so we wouldn't be able to provide the `class_summaries` API.

# Solution
Let's change the API of `class_summaries` into `class_method_signatures` which returns method fully qualified names and their signatures.
In a future diff, we should be able to implement this for Pyrefly.

Reviewed By: tianhan0

Differential Revision: D80697131

fbshipit-source-id: 9b3527625d968ad1f1c250103d4c05465e207c90