Add a generic type to represent a function signature in the PyrePysaAPI interface
Summary:
# Context
During model parsing, we rely on Pyre1's `Type.Callable.t` to represent the signature of a function. This is used to verify that the user model matches with the definition we see. This is essential to find typos in parameter names, for instance.

# Problem
In a follow-up, we will use pyrefly to provide us the information about function signatures.
Requiring a full `Type.Callable.t` is problematic for pyrefly because that would require us to serialize the entire structure of types (unions, type variables, etc.).

# Solution
During model parsing, the only information we need is about the calling convention for parameters, i.e positional/non-positional arguments, *args and **kwargs. In D80936991, we already implemented the serialization of all the data we need.
In this diff, we implement a type `FunctionSignature.t` that will be used as a common representation for a function signature between Pyre1 and Pyrefly.

For now, it still stores a whole `Type.t` for the annotation. This will be addressed in a follow-up.

Reviewed By: tianhan0

Differential Revision: D80951292

fbshipit-source-id: 169bf9064afe9729f902fe8a56c78826688ab196