Add a generic type to represent a type in the PyrePysaAPI interface
Summary:
# Context
We are working on using Pyrefly's type information rather than Pyre1's type information.

# Problem
The model parsing step (and many other steps of Pysa) rely on `Type.t`, the internal representation of a type in Pyre1. However, Pyrefly obviously uses a different internal type representation. We want to avoid having to serialize the data structure representating types in Pyrefly, since this would add a lot of coupling between Pyrefly and Pysa

# Solution
As an internative, we introduce `PysaType.t` which can be used to represent a type from pyre1 or pyrefly, and has a minimal interface - Only what we strictly need in Pysa.

This diff updates the `PyrePysaEnvironment`, the interface for Pyre1 and Pysa, to use `PysaType`. In a follow-up, we will use `PysaType` to represent type information from Pyrefly.

Reviewed By: tianhan0

Differential Revision: D81038995

fbshipit-source-id: 3d86c2604dc18ddf9efa59e600705860bccd68e9