Reset shared memory between tests
Summary:
When running tests with pyrefly, we have one test that fails due to exhaustive the hash table.
This is because the PyreflyAPI for pysa is far from optimal and uses a lot of keys.

We tried to work around this in D81336681 by cleaning up the shared memory after use, but this didn't work: the problem is that we are exhausting the hash table size (using all slots), while the heap size itself is fine. The problem is that the shared memory hash table implementation never removes entries from the hash table (it sets the value to NULL, but keeps the entry).

To work around this problem, this diff proposes to reset the shared memory between tests, which is pretty cheap anyway (cost of `memset(0)` on heap + hash table + set some global variables to 0). This also fixes potential problems with conflicting tests (using the same entry).

There is just one problem due to our interning logic: since we memoize the interned id on first use, we need to add callbacks to explicitly write all memoized interned values to shared memory again. This should be cheap.

Reviewed By: tianhan0

Differential Revision: D81449734

fbshipit-source-id: 88b3f096f04fec58723f6380f07d59090376481b