Run class hierarchy graphs integration tests with pyrefly results
Summary:
This diff introduces a way to run some of our integration tests (for now, just the class hierarchy graph tests) with Pyrefly instead of Pyre1.
This requires running `pyrefly check` on a temporary directory containing all the sources.
Note that this is significantly slower than running regular tests.

To run the tests, one must define the `PYREFLY_BINARY` environment variable to the path of the pyrefly binary, then `make test` or `dune exec <test>`.

Reviewed By: tianhan0

Differential Revision: D81322186

fbshipit-source-id: 858cec6266d3f2fc6802c832f025bc6d0e13a970