Add a cache for projects in pysa tests
Summary:
# Context
Now that we can run some of our tests with pyrefly as a frontend, some of those tests are really slow.
While pyrefly is blazingly fast, it has a high startup cost (presumably because of rust and static linking?). Some of our tests require to type check between 50 and 100 small projects. In addition, the current logic to import results in the Pysa OCaml backend is highly unoptimized. Type checking a small project can take around 1 second.
TL, DR: our tests with pyrefly are really slow (one test takes ~4 minutes)

# Solution
This diff proposes a very basic solution, which is to cache the shared memory to file and use the cache when the source code to analyze is the same. This is very common for model validation and model query tests.

For instance, this makes the interprocedural_analyses/taint/test/modelTest.exe run in 25 seconds instead of 4 minutes on my machine.

# Future
In the future, we could probably do even better, and perform a single type check for all the source code used in tests, and use that result. However, that requires a large refactoring in our tests, so the current solution is probably preferable in the meantime.

Reviewed By: tianhan0

Differential Revision: D81792951

fbshipit-source-id: bae2768fda6c57eaf3d048c5260c7da260e136ac