Split the end to end taint integration tests into different tests
Summary:
# Problem
Running `make test` takes about 2 minutes. The bottleneck is the end-to-end integration tests for Pysa.
That test runs pysa on a list of source files and checks that we find the expected issue(s).
The first problem is that these tests are run sequentially. This is because the test allocates a shared memory heap at the end, and running multiple analysis on the same heap would lead to data races.
The second problem is that one test is significantly slower than the others, `sanitize_tito_shaping.py`. This is apparently due to the sanity check that we enable on tests, via the environment variable PYSA_CHECK_INVARIANTS=1.

# Solution
Let's split these tests into multiple jobs, so that `make test` actually runs those in separate processes.
This requires a small change to the configuration test, so that we produce one dune rule per test.
Then, we can also disable invariant checking for `sanitize_tito_shaping.py` to fix the bottleneck.

Reviewed By: tianhan0

Differential Revision: D81938499

fbshipit-source-id: 8a7d5e46815364babae69b290055fb860b0bc11c